auth

+++ 09_auth/src/modules/recados/recados.service.ts
@@ -15,36 +15,45 @@ export class RecadosService {
   ) { }
 
   findBaseOptions: FindManyOptions<Recado> = {
-    //relations: ['de', 'para'],
+    relations: ['de', 'para'], //! para trazer os dados relacionados
     order: { id: 'desc' },
     select: {
-      //id: true,
-      //texto: true,
+      id: true,
+      texto: true,
+      lido: true,
       de: { id: true, nome: true },
       para: { id: true, nome: true },
     }
   }
 
-  async findAll(page: number, limit: number = 0) {
+  async findAll(userId: number, page: number, limit: number = 0) {
     const take = limit;
     const skip = (page - 1) * limit;
 
+    const para = { id: userId };
+
     if (limit === 0)
-      return await this.recadoRepository.find({ ...this.findBaseOptions });
+      return await this.recadoRepository.find({ ...this.findBaseOptions, where: { para } });
 
-    return await this.recadoRepository.find({ ...this.findBaseOptions, take, skip });
-  }
-
-  async findById(id: number) {
-    const recado = await this.recadoRepository.findOne({ ...this.findBaseOptions, where: { id } });
-    if (!recado) {
-      this.throwNotFoundException();
-    }
+    const recado = await this.recadoRepository.find({ ...this.findBaseOptions, take, skip, where: { para } });
+    console.log('recado', recado);
     return recado;
   }
 
-  async create(recadoDto: CreateRecadoDto) {
-    const de = await this.pessoasService.findOne(recadoDto.deId);
+  async findById(userId: number, id: number) {
+    const recado = await this.recadoRepository.findOne({ ...this.findBaseOptions, where: { id } });
+    if (!recado) {
+      this.throwNotFoundException();
+    } else if (recado.de.id !== userId && recado.para.id !== userId) {
+      console.log(recado);
+      throw new HttpException('Recado não pertence a você', HttpStatus.FORBIDDEN);
+    }
+
+    return recado;
+  }
+
+  async create(deId: number, recadoDto: CreateRecadoDto) {
+    const de = await this.pessoasService.findOne(deId);
     const para = await this.pessoasService.findOne(recadoDto.paraId);
 
     const novoRecado = {
@@ -58,11 +67,18 @@ export class RecadosService {
     return await this.recadoRepository.save(recado);
   }
 
-  async update(id: number, recadoDto: UpdateRecadoDto) {
-    await this.findById(id);
+  async update(userId: number, id: number, recadoDto: UpdateRecadoDto) {
+    const recado = await this.findById(userId, id);
+    if (recado.de.id !== userId && recadoDto?.texto && recado.texto !== recadoDto?.texto)
+      throw new HttpException('Só pode editar a mensagem de recados que você enviou.', HttpStatus.FORBIDDEN);
+
+    const propLido = recadoDto?.lido !== null && recadoDto?.lido !== undefined;
+    if (recado.para.id !== userId && propLido && recado.lido !== recadoDto?.lido)
+      throw new HttpException('Só pode marcado como lido recados que você recebeu.', HttpStatus.FORBIDDEN);
+
     const partialDto = { lido: recadoDto?.lido, texto: recadoDto?.texto }
     await this.recadoRepository.update(id, partialDto);
-    return await this.findById(id);
+    return await this.findById(userId, id);
 
     // const partialDto = { lido: recadoDto?.lido, texto: recadoDto?.texto }
     // const recado = await this.recadoRepository.preload({ id: id, ...partialDto })
@@ -70,8 +86,11 @@ export class RecadosService {
     // return await this.recadoRepository.save(recado);
   }
 
-  async remove(id: number) {
-    const recado = await this.findById(id);
+  async remove(userId: number, id: number) {
+    const recado = await this.findById(userId, id);
+    if (recado.de.id !== userId)
+      throw new HttpException('Só pode deletar recados que você enviou.', HttpStatus.FORBIDDEN);
+
     await this.recadoRepository.remove(recado);
     return recado;
   }
